Переписан процессор мертвого китайца
